Rajout des tests authentification dans le pipeline (cf. dockerfile), modif de l'import de test_api_authentification.py pour que se soit compatible avec l'appel de l'interpréteur python depuiçs la racine du dépôt, mise à jour du fichier test_api_serveur.py avec test_api_serveur_v02.py pour que cela reste compatible avec le pipeline.
